New code for minimum enclosing circle by randomized Welzl algorithm. Split the circle code from Geometry.cpp/hpp to Geometry/Circle.cpp,hpp
